# swagger-spec.json
{
  "openapi": "3.0.0",
  "paths": {
    "/api/v1/users/register": {
      "post": {
        "operationId": "UtilisateurController_register",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Compte créé avec succès"
          },
          "409": {
            "description": "Email déjà utilisé"
          }
        },
        "summary": "Créer un compte utilisateur",
        "tags": [
          "Users"
        ]
      }
    },
    "/api/v1/users/login": {
      "post": {
        "operationId": "UtilisateurController_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginUserDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Connexion réussie, retourne un token"
          },
          "401": {
            "description": "Identifiants invalides"
          }
        },
        "summary": "Se connecter",
        "tags": [
          "Users"
        ]
      }
    },
    "/api/v1/users/verify-account": {
      "post": {
        "operationId": "UtilisateurController_verifyAccount",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerifyAccountDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Compte vérifié avec succès"
          },
          "400": {
            "description": "Token invalide ou expiré"
          }
        },
        "summary": "Vérifier le compte via token reçu par email",
        "tags": [
          "Users"
        ]
      },
      "get": {
        "operationId": "UtilisateurController_verifyAccounts",
        "parameters": [
          {
            "name": "token",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Compte vérifié avec succès"
          },
          "400": {
            "description": "Token invalide ou expiré"
          }
        },
        "summary": "Vérifier le compte via token reçu par email",
        "tags": [
          "Users"
        ]
      }
    },
    "/api/v1/users/resend-verification": {
      "post": {
        "operationId": "UtilisateurController_resendVerification",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ForgotPasswordDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Email renvoyé si le compte existe"
          }
        },
        "summary": "Renvoyer l'email de vérification",
        "tags": [
          "Users"
        ]
      }
    },
    "/api/v1/users/logout/{id}": {
      "post": {
        "operationId": "UtilisateurController_logout",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID de l'utilisateur",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Se déconnecter",
        "tags": [
          "Users"
        ]
      }
    },
    "/api/v1/users/forgot-password": {
      "post": {
        "operationId": "UtilisateurController_forgotPassword",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ForgotPasswordDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Email envoyé si le compte existe"
          }
        },
        "summary": "Demander un lien de réinitialisation du mot de passe",
        "tags": [
          "Users"
        ]
      }
    },
    "/api/v1/users/reset-password": {
      "post": {
        "operationId": "UtilisateurController_resetPassword",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResetPasswordDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Mot de passe réinitialisé"
          },
          "400": {
            "description": "Token invalide ou expiré"
          }
        },
        "summary": "Réinitialiser le mot de passe via token",
        "tags": [
          "Users"
        ]
      },
      "get": {
        "operationId": "UtilisateurController_validateResetToken",
        "parameters": [
          {
            "name": "token",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Valider le token de réinitialisation",
        "tags": [
          "Users"
        ]
      }
    },
    "/api/v1/users/{id}/change-password": {
      "patch": {
        "operationId": "UtilisateurController_changePassword",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID de l'utilisateur",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangePasswordDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Mot de passe modifié"
          },
          "401": {
            "description": "Mot de passe actuel incorrect"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Changer son mot de passe (connecté)",
        "tags": [
          "Users"
        ]
      }
    },
    "/api/v1/users/all": {
      "get": {
        "operationId": "UtilisateurController_findAll",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "example": 20,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Lister tous les utilisateurs (admin)",
        "tags": [
          "Users"
        ]
      }
    },
    "/api/v1/users/{id}": {
      "get": {
        "operationId": "UtilisateurController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID de l'utilisateur",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Utilisateur trouvé"
          },
          "404": {
            "description": "Utilisateur introuvable"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Récupérer un utilisateur par ID",
        "tags": [
          "Users"
        ]
      },
      "patch": {
        "operationId": "UtilisateurController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID de l'utilisateur",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Profil mis à jour"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Mettre à jour le profil",
        "tags": [
          "Users"
        ]
      },
      "delete": {
        "operationId": "UtilisateurController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID de l'utilisateur",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Compte supprimé"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Supprimer un compte (soft delete)",
        "tags": [
          "Users"
        ]
      }
    },
...
  },
  "info": {
    "title": "projet-x-service",
    "description": "API du projet-x-service",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
      "CreateUserDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "description": "Adresse email de l'utilisateur",
            "example": "user@example.com"
          },
          "password": {
            "type": "string",
            "description": "Mot de passe (min 8 caractères)",
            "example": "Password123!",
            "minLength": 8
          },
          "fullName": {
            "type": "string",
            "description": "Nom complet",
            "example": "Kaboré Moussa"
          },
          "language": {
            "type": "string",
            "description": "Langue préférée",
            "example": "fr",
            "default": "fr"
          },
          "timezone": {
            "type": "string",
            "description": "Fuseau horaire",
            "example": "Africa/Abidjan",
            "default": "Africa/Abidjan"
          }
        },
        "required": [
          "email",
          "password",
          "fullName"
        ]
      },
      "LoginUserDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "description": "Email de l'utilisateur",
            "example": "user@example.com"
          },
          "password": {
            "type": "string",
            "description": "Mot de passe",
            "example": "Password123!"
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "VerifyAccountDto": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string",
            "example": "token_recu_par_mail"
          }
        },
        "required": [
          "token"
        ]
      },
      "ForgotPasswordDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "user@example.com"
          }
        },
        "required": [
          "email"
        ]
      },
      "ResetPasswordDto": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string",
            "description": "Token reçu par email"
          },
          "newPassword": {
            "type": "string",
            "description": "Nouveau mot de passe",
            "minLength": 8
          }
        },
        "required": [
          "token",
          "newPassword"
        ]
      },
      "ChangePasswordDto": {
        "type": "object",
        "properties": {
          "currentPassword": {
            "type": "string",
            "description": "Mot de passe actuel",
            "example": "OldPassword123!"
          },
          "newPassword": {
            "type": "string",
            "description": "Nouveau mot de passe",
            "example": "NewPassword456!",
            "minLength": 8
          }
        },
        "required": [
          "currentPassword",
          "newPassword"
        ]
      },
      "UpdateUserDto": {
        "type": "object",
        "properties": {
          "fullName": {
            "type": "string",
            "example": "Kaboré Moussa"
          },
          "bio": {
            "type": "string",
            "example": "Développeur passionné de biographie."
          },
          "avatarUrl": {
            "type": "string",
            "example": "https://cdn.example.com/avatar.jpg"
          },
          "language": {
            "type": "string",
            "example": "fr"
          },
          "timezone": {
            "type": "string",
            "example": "Africa/Abidjan"
          }
        }
      },
...
    }
  }
}